Main GUI
Main GUI and restructure into View
